Updated Modified Constraint

+++ Methods/countlastsecond.java
@@ -2,13 +2,13 @@
 import java.util.Scanner;
 
 public class countlastsecond{
-    static int lastsecond(int n){
+    static long lastsecond(long n){
         if (n<9) return 0;
         return (n/10)%10;
     }
     public static void main(String[] args){
         Scanner sc =new Scanner(System.in);
-        int n = sc.nextInt();
+        long n = sc.nextLong();
         System.out.println(lastsecond(n));
     }
 }
